clinical-conversation-summarization/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# AI/ML Libraries
torch>=2.0.0
transformers==4.35.2
sentence-transformers==2.2.2
# CHANGED: Switched from vosk to openai-whisper
openai-whisper
setuptools-rust
spacy==3.6.1
scispacy==0.5.3

# NLP & Summarization
nltk==3.8.1
rouge-score==0.1.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Utilities
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1

#run on your terminal
#pip install openai-whisper setuptools-rust
#Quick Check: Type ffmpeg -version in your CMD. If you get an error, you'll need to download it from ffmpeg.org and add it to your System PATH.